Create just single PePuRunner on start
Fix an issue with locking by creating just a single PePuRunner on the
process start up in run_pepubot function.

Previously a new PePuRunner was created for each incoming Slack message
and the locking of the shared resources (e.g. runner state and the
lottery box) was done with an instance attribute of the PePuRunner.
Therefore there was actually a different lock for each PePuRunner and
the locking wasn't able to protect two distinct message processing flows
to mess up each other.